Moved Lie groups functional to Labs.
